postgres_to_es/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
asttokens==2.4.1
certifi==2024.2.2
decorator==5.1.1
elastic-transport==8.13.0
elasticsearch==8.13.0
executing==2.0.1
flake8==6.1.0
iniconfig==2.0.0
ipython==8.13.0
jedi==0.19.1
matplotlib-inline==0.1.6
mccabe==0.7.0
packaging==23.2
parso==0.8.4
pexpect==4.9.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg==3.1.18
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pydantic==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
pyflakes==3.1.0
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.2
typing_extensions==4.11.0
urllib3==2.2.1
uWSGI==2.0.24
wcwidth==0.2.13
